+++ source/extensions/omni.isaac.wheeled_robots/premake5.lua
@@ -9,8 +9,8 @@ project_ext( ext, { generate_ext_project=true })
     add_files("omni/isaac/wheeled_robots/nodes", "python/nodes/**.py")
     add_files("omni/isaac/wheeled_robots/robots", "python/robots/**.py")
     add_files("omni/isaac/wheeled_robots/controllers", "python/controllers/**.py")
-
-
+    add_files("omni/isaac/wheeled_robots/tests", "python/tests/**.py")
+    
     add_ogn_dependencies(ogn, {"omni/isaac/wheeled_robots/nodes"})
 
     repo_build.prebuild_copy {
@@ -22,6 +22,7 @@ project_ext( ext, { generate_ext_project=true })
         { "omni/isaac/wheeled_robots/robots", ogn.python_target_path.."/robots" },
         { "docs", ext.target_dir.."/docs" },
         { "data", ext.target_dir.."/data" },
+        { "omni/isaac/wheeled_robots/tests", ogn.python_tests_target_path },
     }