+++ source/extensions/omni.replicator.isaac/docs/CHANGELOG.md
@@ -1,10 +1,22 @@
 # Changelog
 
+## [1.1.2] - 2022-07-26
+### Changed
+- Changed articulation tendon properties write nodes to be sequential.
+### Added
+- Added articulation material properties randomization.
+
+## [1.1.1] - 2022-07-25
+### Fixed
+- Fixed the tests by moving away from using omnigraph bundles and just doing static type resolution on the output of the distribution nodes 
+
 ## [1.1.0] - 2022-07-22
 ### Added
 - Added mass, inertia, material properties, rest offset, and contact offset for rigid prim view randomization
 - Added mass, inertia, and tendon properties for articulation view randomization
 - Added additive and scaling operations for orientation randomization
+- Added pytorch rgb writer with replicator API for isaac gym
+- Added pytorch listener for provide direct access to batched pytorch tensors from gym simulations
 
 ## [1.0.1] - 2022-07-20
 ### Changed